analysis/requirements.txt
attrs==21.4.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
blis==0.7.5
bson==0.5.10
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
cymem==2.0.6
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.2.0/de_core_news_sm-3.2.0-py3-none-any.whl
decorator==5.1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
filelock==3.7.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipython==7.29.0
itemadapter==0.6.0
itemloaders==1.0.4
jedi==0.18.0
Jinja2==3.0.3
jmespath==1.0.0
langcodes==3.2.1
lxml==4.9.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
murmurhash==1.0.6
numpy==1.21.4
packaging==21.2
pandas==1.4.3
parsel==1.6.0
parso==0.8.2
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
preshed==3.0.6
progressbar2==3.55.0
prompt-toolkit==3.0.22
Protego==0.2.1
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
PyDispatcher==2.0.5
Pygments==2.10.0
pymongo==3.12.1
pyOpenSSL==22.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-utils==2.5.6
pytz==2022.1
queuelib==1.6.2
rapidfuzz==1.8.2
regex==2021.11.10
requests==2.26.0
requests-file==1.5.1
Scrapy==2.6.1
service-identity==21.1.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
spaczz==0.5.3
srsly==2.4.2
thinc==8.0.13
tldextract==3.3.0
tqdm==4.62.3
traitlets==5.1.1
Twisted==22.4.0
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
w3lib==1.22.0
wasabi==0.8.2
wcwidth==0.2.5
zope.interface==5.4.0
